mdio-proxy: Add support to access PHY by C45 in latest kernel
In previous kernel, C22 and C45 transactions are mixed over a combined
API calls. But since kernel v6.3-rc1, it has added the dedicated C45
API to the MDIO bus driver, so mdio-proxy.ko will not be applicable to
the kernel since v6.3-rc1 and needs to be updated with the kernel.

Signed-off-by: Wei Fang <wei.fang@nxp.com>